Pull remaining DSL integration tests into dsl module
